Move event loop code to lib
