+++ Cargo.toml
@@ -31,7 +31,7 @@ clap = { version = "2.33.0", optional = true }
 libc = "0.2"
 liquid = "0.22"
 chrono = { version = "0.4.19", optional = true }
-futures = { version = "0.3.12", optional = true }
+futures = "0.3.12"
 slog = { version = "2.5.2", optional = true }
 slog-stdlog = { version = "4.0.0", optional = true }
 slog-scope = { version = "4.3.0", optional = true }
