+++ src/bin/leftwm-worker.rs
@@ -1,24 +1,13 @@
 mod common;
 
 use leftwm::{
-    child_process::{self, Nanny},
+    child_process::{self},
     config::ThemeLoader as _,
-    models::FocusBehaviour,
 };
 
-use leftwm::{
-    CommandPipe, DisplayEvent, DisplayEventHandler, DisplayServer, Manager, Mode, State,
-    StateSocket, Window, Workspace, XlibDisplayServer,
-};
+use leftwm::{DisplayEventHandler, DisplayServer, Manager, XlibDisplayServer};
 use std::panic;
-use std::path::{Path, PathBuf};
-use std::sync::{atomic::Ordering, Arc, Once};
-
-use common::config::Config;
-
-fn get_events<T: DisplayServer<C>, C: leftwm::config::Config>(ds: &mut T) -> Vec<DisplayEvent> {
-    ds.get_next_events()
-}
+use std::sync::Arc;
 
 use slog::{o, Drain};
 
@@ -46,148 +35,18 @@ fn main() {
             config: config.clone(),
         };
 
-        rt.block_on(event_loop(
-            &mut manager,
+        rt.block_on(manager.event_loop(
             &mut display_server,
             &handler,
             config,
-            state,
-            theme_loader,
+            &state,
+            &theme_loader,
         ));
     });
 
     match completed {
         Ok(_) => log::info!("Completed"),
         Err(err) => log::error!("Completed with error: {:?}", err),
-    }
-}
-
-fn place_runtime_file<P>(path: P) -> std::io::Result<PathBuf>
-where
-    P: AsRef<Path>,
-{
-    xdg::BaseDirectories::with_prefix("leftwm")?.place_runtime_file(path)
-}
-
-async fn timeout(mills: u64) {
-    use tokio::time::{sleep, Duration};
-    sleep(Duration::from_millis(mills)).await;
-}
-
-async fn event_loop<CMD>(
-    manager: &mut Manager<CMD>,
-    display_server: &mut XlibDisplayServer<Arc<Config>>,
-    handler: &DisplayEventHandler<Arc<Config>>,
-    config: Arc<Config>,
-    state: common::state::State,
-    theme_loader: common::theme_setting::ThemeLoader,
-) {
-    let socket_file = place_runtime_file("current_state.sock")
-        .expect("ERROR: couldn't create current_state.sock");
-    let mut state_socket = StateSocket::default();
-    state_socket
-        .listen(socket_file)
-        .await
-        .expect("ERROR: couldn't connect to current_state.sock");
-
-    let pipe_file =
-        place_runtime_file("commands.pipe").expect("ERROR: couldn't create commands.pipe");
-    let mut command_pipe = CommandPipe::new(pipe_file)
-        .await
-        .expect("ERROR: couldn't connect to commands.pipe");
-
-    //start the current theme
-    let after_first_loop: Once = Once::new();
-
-    //main event loop
-    let mut event_buffer = vec![];
-    loop {
-        if manager.mode == Mode::Normal {
-            state_socket.write_manager_state(manager).await.ok();
-        }
-        display_server.flush();
-
-        let mut needs_update = false;
-        tokio::select! {
-            _ = display_server.wait_readable(), if event_buffer.is_empty() => {
-                event_buffer.append(&mut get_events(display_server));
-                continue;
-            }
-            //Once in a blue moon we miss the focus event,
-            //This is to double check that we know which window is currently focused
-            _ = timeout(100), if event_buffer.is_empty() && manager.focus_manager.behaviour == FocusBehaviour::Sloppy => {
-                let mut focus_event = display_server.verify_focused_window();
-                event_buffer.append(&mut focus_event);
-                continue;
-            }
-            Some(cmd) = command_pipe.read_command(), if event_buffer.is_empty() => {
-                needs_update = manager.external_command_handler(&state, &config, &theme_loader, cmd) || needs_update;
-                display_server.update_theme_settings(manager.theme_setting.clone());
-            }
-            else => {
-                event_buffer.drain(..).for_each(|event| needs_update = handler.process(manager, &state, event) || needs_update);
-            }
-        }
-
-        //if we need to update the displayed state
-        if needs_update {
-            match &manager.mode {
-                Mode::Normal => {
-                    let windows: Vec<&Window> = manager.windows.iter().collect();
-                    let focused = manager.focused_window();
-                    display_server.update_windows(windows, focused, manager);
-                    let workspaces: Vec<&Workspace> = manager.workspaces.iter().collect();
-                    let focused = manager.focused_workspace();
-                    display_server.update_workspaces(workspaces, focused);
-                }
-                //when (resizing / moving) only deal with the single window
-                Mode::ResizingWindow(h) | Mode::MovingWindow(h) => {
-                    let focused = manager.focused_window();
-                    let windows: Vec<&Window> = (&manager.windows)
-                        .iter()
-                        .filter(|w| &w.handle == h)
-                        .collect();
-                    display_server.update_windows(windows, focused, manager);
-                }
-            }
-        }
-
-        //preform any actions requested by the handler
-        while !manager.actions.is_empty() {
-            if let Some(act) = manager.actions.pop_front() {
-                if let Some(event) = display_server.execute_action(act) {
-                    event_buffer.push(event);
-                }
-            }
-        }
-
-        //after the very first loop run the 'up' scripts (global and theme). we need the unix
-        //socket to already exist.
-        after_first_loop.call_once(|| {
-            match Nanny::run_global_up_script() {
-                Ok(child) => {
-                    child.map(|child| manager.children.insert(child));
-                }
-                Err(err) => log::error!("Global up script faild: {}", err),
-            }
-            match Nanny::boot_current_theme() {
-                Ok(child) => {
-                    child.map(|child| manager.children.insert(child));
-                }
-                Err(err) => log::error!("Theme loading failed: {}", err),
-            }
-
-            state.load(manager);
-        });
-
-        if manager.reap_requested.swap(false, Ordering::SeqCst) {
-            manager.children.reap();
-        }
-
-        if manager.reload_requested {
-            state_socket.shutdown().await;
-            break;
-        }
     }
 }
