+++ src/event_loop.rs
@@ -1,0 +1,146 @@
+use crate::{
+    child_process::Nanny,
+    config::{Config, ThemeLoader},
+    models::FocusBehaviour,
+};
+use crate::{
+    CommandPipe, DisplayEventHandler, DisplayServer, Manager, Mode, State, StateSocket, Window,
+    Workspace, XlibDisplayServer,
+};
+use futures::prelude::*;
+use std::path::{Path, PathBuf};
+use std::sync::{atomic::Ordering, Once};
+
+impl<CMD> Manager<CMD> {
+    pub async fn event_loop<C: Config>(
+        &mut self,
+        display_server: &mut XlibDisplayServer<C>,
+        handler: &DisplayEventHandler<C>,
+        config: C,
+        state: &impl State,
+        theme_loader: &impl ThemeLoader,
+    ) {
+        let socket_file = place_runtime_file("current_state.sock")
+            .expect("ERROR: couldn't create current_state.sock");
+        let mut state_socket = StateSocket::default();
+        state_socket
+            .listen(socket_file)
+            .await
+            .expect("ERROR: couldn't connect to current_state.sock");
+
+        let pipe_file =
+            place_runtime_file("commands.pipe").expect("ERROR: couldn't create commands.pipe");
+        let mut command_pipe = CommandPipe::new(pipe_file)
+            .await
+            .expect("ERROR: couldn't connect to commands.pipe");
+
+        //start the current theme
+        let after_first_loop: Once = Once::new();
+
+        //main event loop
+        let mut event_buffer = vec![];
+        let notify = display_server.task_notify();
+        loop {
+            if self.mode == Mode::Normal {
+                state_socket.write_manager_state(self).await.ok();
+            }
+            display_server.flush();
+
+            let mut needs_update = false;
+            tokio::select! {
+                _ = display_server.wait_readable(), if event_buffer.is_empty() => {
+                    event_buffer.append(&mut get_events(display_server));
+                    continue;
+                }
+                //Once in a blue moon we miss the focus event,
+                //This is to double check that we know which window is currently focused
+                _ = timeout(100), if event_buffer.is_empty() && self.focus_manager.behaviour == FocusBehaviour::Sloppy => {
+                    let mut focus_event = display_server.verify_focused_window();
+                    event_buffer.append(&mut focus_event);
+                    continue;
+                }
+                Some(cmd) = command_pipe.read_command(), if event_buffer.is_empty() => {
+                    needs_update = self.external_command_handler(state, &config, theme_loader, cmd) || needs_update;
+                    display_server.update_theme_settings(self.theme_setting.clone());
+                }
+                else => {
+                    event_buffer.drain(..).for_each(|event| needs_update = handler.process(self, state, event) || needs_update);
+                }
+            }
+
+            //if we need to update the displayed state
+            if needs_update {
+                match &self.mode {
+                    Mode::Normal => {
+                        let windows: Vec<&Window> = self.windows.iter().collect();
+                        let focused = self.focused_window();
+                        display_server.update_windows(windows, focused, self);
+                        let workspaces: Vec<&Workspace> = self.workspaces.iter().collect();
+                        let focused = self.focused_workspace();
+                        display_server.update_workspaces(workspaces, focused);
+                    }
+                    //when (resizing / moving) only deal with the single window
+                    Mode::ResizingWindow(h) | Mode::MovingWindow(h) => {
+                        let focused = self.focused_window();
+                        let windows: Vec<&Window> =
+                            (&self.windows).iter().filter(|w| &w.handle == h).collect();
+                        display_server.update_windows(windows, focused, self);
+                    }
+                }
+            }
+
+            //preform any actions requested by the handler
+            while !self.actions.is_empty() {
+                if let Some(act) = self.actions.pop_front() {
+                    if let Some(event) = display_server.execute_action(act) {
+                        event_buffer.push(event);
+                    }
+                }
+            }
+
+            //after the very first loop run the 'up' scripts (global and theme). we need the unix
+            //socket to already exist.
+            after_first_loop.call_once(|| {
+                match Nanny::run_global_up_script() {
+                    Ok(child) => {
+                        child.map(|child| self.children.insert(child));
+                    }
+                    Err(err) => log::error!("Global up script faild: {}", err),
+                }
+                match Nanny::boot_current_theme() {
+                    Ok(child) => {
+                        child.map(|child| self.children.insert(child));
+                    }
+                    Err(err) => log::error!("Theme loading failed: {}", err),
+                }
+
+                state.load(self);
+            });
+
+            if self.reap_requested.swap(false, Ordering::SeqCst) {
+                self.children.reap();
+            }
+
+            if self.reload_requested {
+                state_socket.shutdown().await;
+                break;
+            }
+        }
+    }
+}
+
+fn place_runtime_file<P>(path: P) -> std::io::Result<PathBuf>
+where
+    P: AsRef<Path>,
+{
+    xdg::BaseDirectories::with_prefix("leftwm")?.place_runtime_file(path)
+}
+
+async fn timeout(mills: u64) {
+    use tokio::time::{sleep, Duration};
+    sleep(Duration::from_millis(mills)).await;
+}
+
+fn get_events<T: DisplayServer<C>, C: Config>(ds: &mut T) -> Vec<crate::DisplayEvent> {
+    ds.get_next_events()
+}
