+++ src/lib.rs
@@ -21,6 +21,7 @@ mod display_action;
 mod display_event;
 pub mod display_servers;
 pub mod errors;
+mod event_loop;
 mod handlers;
 pub mod layouts;
 pub mod models;
